+++ LookupAnything/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* Fixed search UI in 1.53.0 deselecting the search box when the cursor isn't over it.
+
 ## 1.53.0
 Released 26 October 2025 for SMAPI 4.3.1 or later.